Refactors StopOutputMiddleware
