+++ Classes/Middleware/Frontend/StopOutputMiddleware.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Bithost\Pdfviewhelpers\Middleware;
+namespace Bithost\Pdfviewhelpers\Middleware\Frontend;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
@@ -10,14 +10,16 @@ use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use TYPO3\CMS\Core\Http\NullResponse;
 
-class PdfViewHelpersStopOutputMiddleware implements MiddlewareInterface
+class StopOutputMiddleware implements MiddlewareInterface
 {
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $response = $handler->handle($request);
+
         if ($GLOBALS['TSFE']->applicationData['tx_pdfviewhelpers']['pdfOutput'] ?? false) {
             return new NullResponse();
         }
+
         return $response;
     }
 }
